Auto clean (#4)
* Added -r option. New option will backup original_gamelist and use the gamelist.xml as the clean gamelist

* Followed upper-case variable format of existing script.

* Remove extra comments that are redundant by reading code.

* Add check for missing parameters

* Add s to check for readability

* Add example usage

* Update comment and hope eq 0 comes through

* Add check for invalid gamelist.xml

* Remove extra check I added for gamelist. Move 0 filenames check.

* Print input filename if readlink didn't populate original_gamelist

* Add missing space

* Move global variable up to other global variable

* Only need to shift the -r option, there is no value for the key

* Adding message indicating the replace option is being used.

* Added check for no more results incase someone re-runs the process.

* Print file we are working on incase user supplied multple files

* Move printing gamelist to before -r option changes it.

* Add -a|all option.  Use default retropie 4.2 gamelist location.

* Rename variable as it was already used

* Removing quotes as getting cd error

* Extract full_path logic to reuse

* Fix missing check on if statement

* Add missing $

* Put while loop back with < <, could not get output from xmlstarlet to store in variable with new lines.

* Fix function to use temp variable as return value

* Remove return

* Try to fix the path issue with -a

* Fixing typo

* Remove attempt to reuse code that wasn't needed.

* Get full path to gamelist.xml files

* Trying to force the * to expand into directories

* Trying without quotes

* Forgot to put the  back instead of

* Add two lines after file size difference for -a option:

* Update options help message.  Add -l|list option to specify the gamelists root folder.

* Remove copy-paste error.

* Update comments

* Add -e|eliminate option to remove any old backups when a user is ready to remove them.

* Adding eliminate message and fixing E in Echo

* Cleaning up stdout.

+++ gamelist-cleaner.sh
@@ -9,9 +9,16 @@
 # Run the script with '--help' to get more info.
 # 
 # meleu - 2017/Jun
+# kaltinril - 2017-08-19 - Added -r option to replace the existing gameslist
 
+# Global Variables
+REPLACE_GAMELIST=false
+DO_ALL=false
+LISTS_DIR="$HOME/.emulationstation/gamelists"
 ROMS_DIR="$HOME/RetroPie/roms"
+ELIMINATE_BACKUPS=false
 
+# Read only Variables
 readonly SCRIPT_DIR="$(dirname "$0")"
 readonly SCRIPT_NAME="$(basename "$0")"
 readonly SCRIPT_FULL="$SCRIPT_DIR/$SCRIPT_NAME"
@@ -21,8 +28,12 @@ readonly USAGE="Usage:
 $0 [OPTIONS] [gamelist.xml]...
 "
 
+readonly EXAMPLE="Example:
+$0 ~/.emulationstation/gamelists/nes/gamelist.xml
+"
+
 readonly HELP="
-This script gets a gamelist.xml as input and check if the path for the games
+This script gets a gamelist.xml as input and checks if the path for the games
 leads to an existing file. If the file doesn't exist, the <game> entry will
 be deleted and a cleaner gamelist.xml file will be generated.
 
@@ -30,6 +41,7 @@ The resulting file will be named \"gamelist.xml-clean\" and will be in the
 same folder as the original file. Nothing changes in the original gamelist.xml.
 
 $USAGE
+$EXAMPLE
 The OPTIONS are:
 
 -h|--help           print this message and exit.
@@ -42,6 +54,13 @@ The OPTIONS are:
 
 -d|--directory DIR  specifies the ROMs directory. Default:
                     $ROMS_DIR
+
+-r|--replace        Force replace the gamelist.xml file (Creates backup of original)
+
+-a|--all            Automatically clean all gamelists that exist in lists folder
+
+-l|--gamelist DIR   specifies the gamelist directory.  Default:
+                    $LISTS_DIR
 "
 
 function update_script() {
@@ -71,6 +90,13 @@ function update_script() {
     exit 0
 }
 
+function eliminate_backup_files() {
+    backups=$(ls ${LISTS_DIR}/*/gamelist.xml-orig*)
+    for file in $backups; do
+      echo "Removing: $file"
+      rm $file
+    done
+}
 
 while [[ -n "$1" ]]; do
     case "$1" in
@@ -91,8 +117,31 @@ while [[ -n "$1" ]]; do
             ROMS_DIR="$1"
             shift
             ;;
+        -r|--replace)
+            shift
+            REPLACE_GAMELIST=true
+            echo "Using replace option"
+            echo
+            ;;
+        -a|--all)
+            shift
+            DO_ALL=true
+            echo "Cleaning all gamelists!"
+            echo
+            ;;
+        -l|--list)
+            shift
+            LISTS_DIR="$1"
+            shift
+            ;;
+        -e|--eliminate)
+            shift
+            ELIMINATE_BACKUPS=true
+            echo "Using eliminate option"
+            echo
+            ;;
         '')
-            echo "ERROR: missing gamelist.xml" >&2
+            echo "ERROR: missing gamelist.xml parameter" >&2
             echo "$HELP" >&2
             exit 1
             ;;
@@ -107,12 +156,27 @@ while [[ -n "$1" ]]; do
     esac
 done
 
-for file in "$@"; do
+# Verify there is at least 1 file after all parameters
+if [ "$#" -eq 0 ] && [ "$DO_ALL" = false ]; then
+    echo "ERROR: missing gamelist.xml parameter" >&2
+    echo "$HELP" >&2
+    exit 1
+fi
+
+# Get list of files to use
+gamelist_files="$@"
+if [ "$DO_ALL" = true ]; then
+    gamelist_files=$(ls ${LISTS_DIR}/*/gamelist.xml)
+fi
+
+for file in $gamelist_files; do
     original_gamelist="$(readlink -e "$file")"
     clean_gamelist="${original_gamelist}-clean"
     gamelist_dir="$(dirname "$original_gamelist")"
+    backup_gamelist="${original_gamelist}-orig-$(date +%s)"
 
     if [[ ! -s "$original_gamelist" ]]; then
+        [[ -z "$original_gamelist" ]] && original_gamelist="$file" # Make sure we print the name if the readlink failed to find a file
         echo "\"$original_gamelist\": file not found or is zero-length. Ignoring..."
         continue
     fi
@@ -126,7 +190,28 @@ for file in "$@"; do
         continue
     fi
 
-    cat "$original_gamelist" > "$clean_gamelist"
+    # What file are we working on?
+    echo "Working on: ${original_gamelist}"
+    
+    # Make backup of gamelist and replace original
+    if [ "$REPLACE_GAMELIST" = true ]; then
+      cat "$original_gamelist" > "$backup_gamelist"
+      clean_gamelist="$original_gamelist"
+      original_gamelist="$backup_gamelist"
+    else
+      # Leave original alone, create a clean separate file
+      cat "$original_gamelist" > "$clean_gamelist"
+    fi
+    
+    # Check to see if we have any entires before we try to loop over them.
+    xml_entries=$(xmlstarlet sel -t -v "/gameList/game/path" "$original_gamelist")
+    if [[ -z $xml_entries ]]; then
+        echo "No entries found, file is empty."
+        echo
+        echo
+        continue
+    fi
+    
     while read -r path; do
         full_path="$path"
         [[ "$path" == ./* ]] && full_path="$ROMS_DIR/$system/$path"
@@ -141,4 +226,11 @@ for file in "$@"; do
     echo
     echo "See the difference between file sizes:"
     du -bh "$original_gamelist" "$clean_gamelist"
+    echo
+    echo
 done
+
+if [ "$ELIMINATE_BACKUPS" = true ];then
+    echo "Removing backups...."
+    eliminate_backup_files
+fi
